+++ pkg/kt_errors/common.go
@@ -1,0 +1,8 @@
+package kt_errors
+
+import "github.com/keytiles/lib-errorhandling-golang/v2/pkg"
+
+const (
+	// Name of package within the lib - used as a prefix for kt_errors.Fault sources and Logging
+	PACKAGE_NAME = pkg.LIB_NAME + ".kt_errors"
+)